Move Chimera and Pegasus elimination orders from generators to algorithms
