+++ dwave_networkx/generators/pegasus.py
@@ -270,51 +270,6 @@ def pegasus_graph(m, create_using=None, node_list=None, edge_list=None, data=Tru
     return G
 
 
-def pegasus_elimination_order(n, coordinates=False):
-    '''
-    Produces a variable elimination order for a pegasus P(n) graph, which provides an upper bound on the treewidth.
-
-    Simple pegasus variable elimination order rules:
-       - eliminate vertical qubits, one column at a time
-       - eliminate horizontal qubits in each column once their adjacent vertical qubits have been eliminated
-
-    Many other orderings are possible giving the same treewidth upper bound of 12n-4 for a P(n)
-    see pegasus_var_order_extra.py for a few.
-    P(n) contains cliques of size 12n-10 so we know the treewidth of P(n) is in [12n-11,12n-4].
-
-    the treewidth bound generated by a variable elimination order can be verified using
-    dwave_networkx.elimination_order_width. Example:
-    import dwave_networkx as dwnx
-    n = 6
-    P = dwnx.generators.pegasus_graph(n)
-    order = pegasus_var_order(n)
-    print(dwnx.elimination_order_width(P,order))
-    '''
-    m = n
-    l = 12
-
-    # ordering for horizontal qubits in each tile, from east to west:
-    h_order = [4, 5, 6, 7, 0, 1, 2, 3, 8, 9, 10, 11]
-    order = []
-    for n_i in range(n):  # for each tile offset
-        # eliminate vertical qubits:
-        for l_i in range(0, l, 2):
-            for l_v in range(l_i, l_i + 2):
-                for m_i in range(m - 1):  # for each column
-                    order.append((0, n_i, l_v, m_i))
-            # eliminate horizontal qubits:
-            if n_i > 0 and not(l_i % 4):
-                # a new set of horizontal qubits have had all their neighbouring vertical qubits eliminated.
-                for m_i in range(m):
-                    for l_h in range(h_order[l_i], h_order[l_i] + 4):
-                        order.append((1, m_i, l_h, n_i - 1))
-
-    if coordinates:
-        return order
-    else:
-        return pegasus_coordinates(n).ints(order)
-
-
 def get_tuple_fragmentation_fn(pegasus_graph):
     """
     Returns a fragmentation function that is specific to pegasus_graph. This fragmentation function,
